backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase==2.18.1
redis==6.4.0
python-dotenv==1.1.1
pydantic-settings==2.11.0
httpx==0.28.1
youtube-transcript-api==1.2.3
langchain-core==1.2.6
langgraph==0.6.7
langchain-openai==1.0.2
langchain-google-genai==3.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-huggingface==1.2.0
langchain-pinecone==0.2.13
pinecone==7.3.0
